Flag to use new PyArrow Dataset code path
PyArrow has a new "experimental" Dataset API, which ParquetDataset
instances if the `use_legacy_dataset` constructor argument is set to
False.  This allows non-partition-key columns to be used to filter rows
via the `filters`, in addition to allowing different partitioning
schemes.

Add a new parameter to `petastorm.reader.make_batch_reader` that
determines whether this new PyArrow Dataset API is used by the
ParquetDataset instance that the Petastorm readers wrap. The parameter
is called `use_experimental_pyarrow_api`, and defaults to False. If
a caller passes True, then the new API will be used; if not, then the
behavior remains unchanged.

Add the same parameter to the `Reader` class, again defaulting to False.
Negate this parameter and use the negated value as the value for
ParquetDataset's `use_legacy_dataset` constructor argument.